@@ -27,18 +27,18 @@
   "dependencies": {
     "dayjs": "^1.10.5",
     "lodash": "^4.17.21",
-    "mobx": "^6.3.2",
-    "mobx-react": "^7.2.0",
-    "react": "^17.0.0",
+    "mobx": "^6.7.0",
+    "mobx-react": "^7.6.0",
     "react-desktop": "^0.3.9",
-    "react-dom": "^17.0.0",
-    "react-transition-group": "^4.4.2"
+    "react-transition-group": "^4.4.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@types/lodash": "^4.14.170",
     "@types/node": "^15.12.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
+    "@types/react": "^18.0.27",
+    "@types/react-dom": "^18.0.10",
     "@types/react-transition-group": "^4.4.1",
     "@typescript-eslint/eslint-plugin": "^5.33.0",
     "@typescript-eslint/parser": "^5.33.0",
@@ -53,7 +53,7 @@
     "stylelint": "^13.13.1",
     "stylelint-config-recess-order": "^2.4.0",
     "stylelint-config-standard": "^22.0.0",
-    "typescript": "^4.1.2",
+    "typescript": "^4.9.4",
     "vite": "^2.6.14",
     "vite-plugin-imagemin": "^0.6.1"
   },
