@@ -1,11 +1,11 @@
-import { StrictMode } from 'react'
-import ReactDOM from 'react-dom'
+import React from 'react'
+import ReactDOM from 'react-dom/client'
 import './styles/index.scss'
 import App from './App'
 
-ReactDOM.render(
-  <StrictMode>
+const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement)
+root.render(
+  <React.StrictMode>
     <App />
-  </StrictMode>,
-  document.getElementById('root')
+  </React.StrictMode>
 )
