@@ -1,15 +1,17 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-declare module "react-desktop/macOs" {
-  export const View: any;
-  export const Radio: any;
-  export const TitleBar: any;
-  export const Toolbar: any;
-  export const Text: any;
-  export const Box: any;
-  export const ListView: any;
-  export const ListViewRow: any;
-  export const Window: any;
-  export const Dialog: any;
-  export const Button: any;
-  export const Checkbox: any;
+
+declare module 'react-desktop/macOs' {
+  export const View: any
+  export const Radio: any
+  export const TitleBar: any
+  export const Toolbar: any
+  export const Text: any
+  export const Box: any
+  export const ListView: any
+  export const ListViewRow: any
+  export const Window: any
+  export const Dialog: any
+  export const Button: any
+  export const Checkbox: any
+  export const ProgressCircle: any
 }
